+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.5.2
+## Added
+- Added support for Recommendations widget display modes (slider (default) or product list); configure via the widget configuration’s `settings.type` property (`slider` or `list`).
+## Fixed
+- Recommendations page works as expected again
+
 ## 1.5.1
 ## Added
 - Support for multiple slider titles on the same page via additional `header` property inside `requestOptions` entries